+++ CHANGELOG.md
@@ -5,12 +5,20 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic
 Versioning](https://semver.org/spec/v2.0.0.html).
 
-## Unreleased
+## v0.5.0 - UNRELEASED
+
+### Added
+
+- Support multiple prefixes in `Nosedrum.Invoker.Split`
 
 ### Changed
 
-- Updated nostrum to 0.6.1.
+- Updated nostrum to 0.7.0.
 - Updated all other dependencies.
+
+### Removed
+
+- Remove member lookup by nickname
 
 ### Fixed